Add folder "Jacoco" and edits
